requirements.txt
altgraph==0.17.4
blinker==1.8.2
click==8.1.7
et_xmlfile==2.0.0
Flask==3.0.3
imagetk==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
ttkbootstrap==1.10.1
tzdata==2024.2
Werkzeug==3.0.6
